# micro-productos/micro-productos-service-api/src/main/resources/swagger/openapi-v3.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    CRUD de producto y tipos
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: https://productstore3.swagger.io/api/v3
  
tags:
  - name: products
    description: Everything about your products
    externalDocs:
      description: Find out more
      url: http://swagger.io
      
paths:
  /producto:
   
    post:
      tags:
        - producto
      summary: Adiciona un nuevo producto
      description: Adiciona un nuevo producto
      operationId: crearProducto
      requestBody:
        description: crea un nuevo producto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoVo'
          application/xml:
            schema:
              $ref: '#/components/schemas/ProductoVo'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProductoVo'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoVo'          
            application/xml:
              schema:
                $ref: '#/components/schemas/ProductoVo'
        '405':
          description: Invalid input

    get:
      tags:
        - producto
      summary: COnsulta todos los productos
      description: COnsulta todos los productos
      operationId: consultaProductos
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoVo'          
            application/xml:
              schema:
                $ref: '#/components/schemas/ProductoVo'
        '404':
          description: producto no encontrados
        '405':
          description: excepcion
    
    put:
      tags:
        - producto
      summary: Actualiza un producto existente
      description: Actualiza un producto existente por id
      operationId: actualizaProducto
      requestBody:
        description: Actualiza un producto existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoVo'
          application/xml:
            schema:
              $ref: '#/components/schemas/ProductoVo'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProductoVo'
        required: true
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoVo'          
            application/xml:
              schema:
                $ref: '#/components/schemas/ProductoVo'
        '400':
          description: Id no valido
        '404':
          description: Producto no encontrado
        '405':
          description: Excepcion


  /producto/findByStatus:
    get:
      tags:
        - producto
      summary: Encuentra producto por tipo
      description: Encuentra producto por tipo
      operationId: consultaProductoPorTipo
      parameters:
        - name: tipoProducto
          in: query
          description: filtra por tipo producto
          required: false
          explode: true
          schema:
            type: string
            default: accesorios
            enum:
              - accesorios
              - blusas
              - jeans
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoVo'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoVo'
        '400':
          description: Tipo no valido

  /producto/{id}:
    get:
      tags:
        - producto
      summary: obtiene producto por id
      description: retorna un producto de acuerdo al id
      operationId: consultaProductoId
      parameters:
        - name: id
          in: path
          description: Id de producto a retornar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoVo'          
            application/xml:
              schema:
                $ref: '#/components/schemas/ProductoVo'
        '400':
          description: Id no valido
        '404':
          description: Producto no encontrado

components:
  schemas:
    ProductoVo:
      required:
        - nombreProducto
        - tipoProducto
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        nombreProducto:
          type: string
          example: collar
        unidades:
          type: integer
          example: 2
        precioVenta:
          type: string
          format: Float
          example: 8
        tipoProducto:
          type: string
          description: tipo de producto 
          enum:
            - accesorios
            - blusas
            - jeans
        fechaCreacion:
          type: string
          format: date-time
          example: 20230528
        fechaModificacion:
          type: string
          format: date-time
          example: 2023-05-28